+++ CHANGELOG.md
@@ -1,13 +1,33 @@
-# [Unreleased-main] - 2026-07-09
+# [Unreleased-main] - 2026-07-10
+
+## Added
+- xxx
+
+## Changed
+- xxx
+
+## Fixed
+- xxx
+
+
+# [0.1.20.1] - 2026-07-10
 
 ## Added
 - xxx
 
 ## Changed
 - Code generalization in asset sizing mixin
+- Cater for a pipe price catalogue per pipe via MesureGroup
+- Removing the units in the KPI labels
+- Energy production (yearly averaged) KPI is updated so that the name is displayed instead of the id
+- Updated tolerance value used in DTK post processing
 
 ## Fixed
+- Check for allowing a user defined lower pipe DN size via measures/templates
+- Bug fix: In __override_pipe_classes_dicts a min_size_idx=0 was not catered for
+- Bug fix: In update_pipe_class_costs() looping over items in enumerate(pipe_classes) while deleting items in pipe_classes resulted in items being skipped
 - Bug: Only update the aggregation count for an ATES and a GeothermalSource in DTK post processing
+
 
 # [0.1.20] - 2026-06-15